backend/services/piper-tts/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
piper-tts>=1.2.0
numpy>=1.24.0
python-multipart>=0.0.6
